// source code/frontend-vite/src/types/cartTypes.ts
import type { Product, PersonalizedGift } from "./productTypes";
import type { User } from "./userTypes";

export interface CartItem {
  id: number;
  cart?: Cart;
  product: Product;
  personalizedGift?: PersonalizedGift | null;
  quantity: number;
  mrpPrice: number;
  sellingPrice: number;
  userId: number;
}

export interface Cart {
  id: number;
  user: User;
  cartItems: CartItem[];
  totalSellingPrice: number;
  totalItem: number;
  totalMrpPrice: number;
  discount: number;
  couponCode: string | null;
}